+++ sayaya-ui/src/net/sayaya/ui/regacy/graph/GraphBox.java
@@ -17,8 +17,8 @@ public class GraphBox<C, N extends Number> extends Graph<Map<C, net.sayaya.ui.re
 	private final AxisContinuous<N> axisY;
 	private String borderColor;
 	private double borderWidth = 1;
-	private double betweenMargin = 20;
-	private double withinMargin = 5;
+	private int betweenMargin = 20;
+	private int withinMargin = 5;
 	private HashMap<C, Map<Integer, BoxAnimated>> shapes = new HashMap<C, Map<Integer, BoxAnimated>>();
 	
 	public GraphBox(int width, int height, AxisDiscretized<C> axisX, AxisContinuous<N> axisY) {
@@ -35,21 +35,21 @@ public class GraphBox<C, N extends Number> extends Graph<Map<C, net.sayaya.ui.re
 		double q3 = axisY.parse(box.getQ3());
 		double max = axisY.parse(box.getMax());
 
-		double minMap = axisY.getY() - min;
-		double q1Map = axisY.getY() - q1;
-		double medMap = axisY.getY() - med;
-		double q3Map = axisY.getY() - q3;
-		double maxMap = axisY.getY() - max;
-		Box<Double> nextBox = new Box<Double>().setMin(minMap).setQ1(q1Map).setMedian(medMap).setQ3(q3Map).setMax(maxMap);
+		int minMap = (int)(axisY.getY() - min);
+		int q1Map = (int)(axisY.getY() - q1);
+		int medMap = (int)(axisY.getY() - med);
+		int q3Map = (int)(axisY.getY() - q3);
+		int maxMap = (int)(axisY.getY() - max);
+		Box<Integer> nextBox = new Box<Integer>().setMin(minMap).setQ1(q1Map).setMedian(medMap).setQ3(q3Map).setMax(maxMap);
 		if(!shapes.containsKey(category) || !shapes.get(category).containsKey(idx)) {
-			double bw = getBoxWidth();
+			int bw = (int)getBoxWidth();
 			Agenda agenda = getAgenda()[idx];
 			double delta = betweenMargin/2.0 + idx*(bw + withinMargin);
-			double x = axisX.parse(category) + delta;
-			Box<Double> initBox = new Box<Double>().setMin(medMap).setQ1(medMap).setMedian(medMap).setQ3(medMap).setMax(medMap);
+			int x = (int) (axisX.parse(category) + delta);
+			Box<Integer> initBox = new Box<Integer>().setMin(medMap).setQ1(medMap).setMedian(medMap).setQ3(medMap).setMax(medMap);
 			BoxAnimated shape = (BoxAnimated) createBox(bw, initBox, agenda)
 				.setBoxNext(nextBox)
-				.setPointNext(new Point<Double, Double> (x, axisY.getY()))
+				.setPointNext(new Point<Integer, Integer> (x, axisY.getY()))
 				.setColor(agenda.getColor())
 				.setBorderWidth(borderWidth).setBorderColor(borderColor)
 				.setX(x).setY(axisY.getY())
@@ -69,7 +69,7 @@ public class GraphBox<C, N extends Number> extends Graph<Map<C, net.sayaya.ui.re
 	protected void setShapes() {
 		shapes.clear();
 		Map<C, Box<N>>[] values = getValue();
-		double bw = getBoxWidth();
+		int bw = (int)getBoxWidth();
 		for(int i = 0; i < values.length; ++i) {
 			Map<C, Box<N>> item = values[i];
 			Agenda agenda = getAgenda()[i];
@@ -77,23 +77,23 @@ public class GraphBox<C, N extends Number> extends Graph<Map<C, net.sayaya.ui.re
 			for(Entry<C, Box<N>> value: item.entrySet()) {
 				C category = value.getKey();
 				Box<N> box = value.getValue();
-				double x = axisX.parse(category) + delta;
+				int x = (int) (axisX.parse(category) + delta);
 				double min = axisY.parse(box.getMin());
 				double q1 = axisY.parse(box.getQ1());
 				double med = axisY.parse(box.getMedian());
 				double q3 = axisY.parse(box.getQ3());
 				double max = axisY.parse(box.getMax());
 				
-				double minMap = axisY.getY() - min;
-				double q1Map = axisY.getY() - q1;
-				double medMap = axisY.getY() - med;
-				double q3Map = axisY.getY() - q3;
-				double maxMap = axisY.getY() - max;
-				Box<Double> initBox = new Box<Double>().setMin(medMap).setQ1(medMap).setMedian(medMap).setQ3(medMap).setMax(medMap);
-				Box<Double> nextBox = new Box<Double>().setMin(minMap).setQ1(q1Map).setMedian(medMap).setQ3(q3Map).setMax(maxMap);
+				int minMap = (int)(axisY.getY() - min);
+				int q1Map = (int)(axisY.getY() - q1);
+				int medMap = (int)(axisY.getY() - med);
+				int q3Map = (int)(axisY.getY() - q3);
+				int maxMap = (int)(axisY.getY() - max);
+				Box<Integer> initBox = new Box<Integer>().setMin(medMap).setQ1(medMap).setMedian(medMap).setQ3(medMap).setMax(medMap);
+				Box<Integer> nextBox = new Box<Integer>().setMin(minMap).setQ1(q1Map).setMedian(medMap).setQ3(q3Map).setMax(maxMap);
 				BoxAnimated shape = (BoxAnimated) createBox(bw, initBox, agenda)
 					.setBoxNext(nextBox)
-					.setPointNext(new Point<Double, Double> (x, axisY.getY()))
+					.setPointNext(new Point<Integer, Integer> (x, axisY.getY()))
 					.setColor(agenda.getColor())
 					.setBorderWidth(borderWidth).setBorderColor(borderColor)
 					.setX(x).setY(axisY.getY())
@@ -111,20 +111,20 @@ public class GraphBox<C, N extends Number> extends Graph<Map<C, net.sayaya.ui.re
 	public AxisContinuous<N> getAxisY() {
 		return axisY;
 	}
-	protected BoxAnimated createBox(double width, Box<Double> box, Agenda agenda) {
+	protected BoxAnimated createBox(int width, Box<Integer> box, Agenda agenda) {
 		return new BoxAnimated(width, box);
 	}
-	public double getBetweenMargin() {
+	public int getBetweenMargin() {
 		return betweenMargin;
 	}
-	public GraphBox<C, N> setBetweenMargin(double betweenMargin) {
+	public GraphBox<C, N> setBetweenMargin(int betweenMargin) {
 		this.betweenMargin = betweenMargin;
 		return this;
 	}
-	public double getWithinMargin() {
+	public int getWithinMargin() {
 		return withinMargin;
 	}
-	public GraphBox<C, N> setWithinMargin(double withinMargin) {
+	public GraphBox<C, N> setWithinMargin(int withinMargin) {
 		this.withinMargin = withinMargin;
 		return this;
 	}
@@ -199,24 +199,24 @@ public class GraphBox<C, N extends Number> extends Graph<Map<C, net.sayaya.ui.re
 		}
 	}
 	public static class BoxAnimated extends ShapeInstance<BoxAnimated> implements HasStroke {
-		private double width;
-		private final Box<Double> box = new Box<Double>();
-		private Box<Double> boxPrev;
-		private Box<Double> boxNext;
+		private int width;
+		private final Box<Integer> box = new Box<Integer>();
+		private Box<Integer> boxPrev;
+		private Box<Integer> boxNext;
 		
-		private Point<Double, Double> pointNext = new Point<Double, Double>(0.0, 0.0);
+		private Point<Integer, Integer> pointNext = new Point<Integer, Integer>(0, 0);
 		private double rotateNext = 0;
-		private Point<Double, Double> pointPrev = new Point<Double, Double>(0.0, 0.0);
+		private Point<Integer, Integer> pointPrev = new Point<Integer, Integer>(0, 0);
 		// private double rotatePrev = 0.0;
 		
 		private String color;
 		private String borderColor;
 		private double borderWidth = 1;
-		private BoxAnimated(double width, Box<Double> box) {
+		private BoxAnimated(int width, Box<Integer> box) {
 			this.width = width;
 			this.box.setMin(box.getMin()).setQ1(box.getQ1()).setMedian(box.getMedian()).setQ3(box.getQ3()).setMax(box.getMax());
 		}
-		public BoxAnimated setBoxNext(Box<Double> box) {
+		public BoxAnimated setBoxNext(Box<Integer> box) {
 			this.boxNext = box;
 			return this;
 		}
@@ -227,7 +227,7 @@ public class GraphBox<C, N extends Number> extends Graph<Map<C, net.sayaya.ui.re
 				pointPrev.setX(getX());
 				pointPrev.setY(getY());
 				// rotatePrev = getRotate();
-				boxPrev = new Box<Double>().setMin(box.getMin()).setQ1(box.getQ1()).setMedian(box.getMedian()).setQ3(box.getQ3()).setMax(box.getMax());
+				boxPrev = new Box<Integer>().setMin(box.getMin()).setQ1(box.getQ1()).setMedian(box.getMedian()).setQ3(box.getQ3()).setMax(box.getMax());
 			}
 			context.setLineWidth(borderWidth);
 			context.scale(1.0, -1.0);
@@ -262,15 +262,16 @@ public class GraphBox<C, N extends Number> extends Graph<Map<C, net.sayaya.ui.re
 			}
 
 			
-			this.box.setMin(box.getMin()).setQ1(box.getQ1()).setMedian(box.getMedian()).setQ3(box.getQ3()).setMax(box.getMax());
+			this.box.setMin(box.getMin().intValue()).setQ1(box.getQ1().intValue()).setMedian(box.getMedian().intValue())
+			.setQ3(box.getQ3().intValue()).setMax(box.getMax().intValue());
 		}
 		
 		@Override
 		public boolean checkIn(double x, double y) {
-			double lx = 0;
-			double rx = width;
-			double ty = box.getQ1();
-			double by = box.getQ3();
+			int lx = 0;
+			int rx = width;
+			int ty = box.getQ1();
+			int by = box.getQ3();
 			
 			if(lx > width-x || rx < width-x) return false;
 			if(ty > -y || by < -y) return false;
@@ -289,10 +290,10 @@ public class GraphBox<C, N extends Number> extends Graph<Map<C, net.sayaya.ui.re
 			return new Box<Double>().setMin(min).setQ1(q1).setMedian(med).setQ3(q3).setMax(max);
 		}
 		
-		public Point<Double, Double> getPointNext() {
+		public Point<Integer, Integer> getPointNext() {
 			return pointNext;
 		}
-		public BoxAnimated setPointNext(Point<Double, Double> pointNext) {
+		public BoxAnimated setPointNext(Point<Integer, Integer> pointNext) {
 			this.pointNext = pointNext;
 			return this;
 		}
